requirements.txt
streamlit
openai
faiss-cpu
numpy
sentence-transformers
python-dotenv